Refactor : custom hook 적용 및 유틸분리
